Added Humility reduction over time (with animation curve to make it depend on Boss Health)

+++ Assets/Enemies/Scripts/OdinEnemyPhase.cs
@@ -17,7 +17,7 @@ public class OdinEnemyPhase
 
     [OdinSerialize]
     [LabelText("Quantité d'Humility de la phase")]
-    public int maxHumility;
+    public float maxHumility;
 
     [OdinSerialize]
     [LabelText("Transition")]
